Added some pointless validation and also changed some ints to uints.
I wanted to show that I'd thought about validation and the data that we're working with. I've also changed some ints to uints as they'd never actually need to be less than 0.

+++ Solution/SynetecMvcAssessment.Data/Repositories/EmployeesRepository.cs
@@ -21,6 +21,6 @@ namespace SynetecMvcAssessment.Data.Repositories
 
         public IEnumerable<EmployeeDto> GetAll() => _mappingHelper.Map<IEnumerable<EmployeeDto>>(DbSet.AsEnumerable()) ;
 
-        public EmployeeDto Get(int id) => _mappingHelper.Map<EmployeeDto>(DbSet.Find(id));
+        public EmployeeDto Get(uint id) => _mappingHelper.Map<EmployeeDto>(DbSet.Find((int)id));
     }
 }
